refactor code for dashboard
